fraud_predictor/requirements.txt
kafka-python
pandas
joblib
scikit-learn
xgboost